requirements.txt
playwright
requests
openai
Pillow